# docs/thornode-openapi.yaml
openapi: 3.0.0
info:
  title: Thornode API
  version: 3.14.1
  contact:
    email: devs@thorchain.org
  description: Thornode REST API.

########################################################################################
# Paths
########################################################################################
paths:
...
  /thorchain/pools:
    parameters:
      - $ref: "#/components/parameters/queryHeight"
    get:
      description: Returns the pool information for all assets.
      operationId: pools
      tags:
        - Pools
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolsResponse"
...
  /thorchain/network:
    parameters:
      - $ref: "#/components/parameters/queryHeight"
    get:
      description: Returns network overview statistics.
      operationId: network
      tags:
        - Network
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkResponse"
...
components:
...
  parameters:
    queryHeight:
      name: height
      in: query
      description: optional block height, defaults to current tip
      required: false
      schema:
        type: integer
        format: int64
        minimum: 0
...
  schemas:
    Pool:
      type: object
      required:
        - asset
        - status
        - pending_inbound_asset
        - pending_inbound_rune
        - balance_asset
        - balance_rune
        - asset_tor_price
        - pool_units
        - LP_units
        - synth_units
        - synth_supply
        - savers_depth
        - savers_units
        - savers_fill_bps
        - savers_capacity_remaining
        - synth_mint_paused
        - synth_supply_remaining
        - derived_depth_bps
      properties:
        asset:
          type: string
          example: "BTC.BTC"
        short_code:
          type: string
          example: "b"
        status:
          type: string
          example: "Available"
        decimals:
          type: integer
          format: int64
          example: 6
        pending_inbound_asset:
          type: string
          example: "101713319"
        pending_inbound_rune:
          type: string
          example: "464993836"
        balance_asset:
          type: string
          example: "3197744873"
        balance_rune:
          type: string
          example: "13460619152985"
        asset_tor_price:
          type: string
          example: "123456"
          description: the USD (TOR) price of the asset in 1e8
        pool_units:
          type: string
          description: the total pool units, this is the sum of LP and synth units
          example: "14694928607473"
        LP_units:
          type: string
          description: the total pool liquidity provider units
          example: "14694928607473"
        synth_units:
          type: string
          description: the total synth units in the pool
          example: "0"
        synth_supply:
          type: string
          description: the total supply of synths for the asset
          example: "0"
        savers_depth:
          type: string
          example: "199998"
          description: the balance of L1 asset deposited into the Savers Vault
        savers_units:
          type: string
          example: "199998"
          description: the number of units owned by Savers
        savers_fill_bps:
          type: string
          example: "4500"
          description: the filled savers capacity in basis points, 4500/10000 = 45%
        savers_capacity_remaining:
          type: string
          example: "1000"
          description: amount of remaining capacity in asset
        synth_mint_paused:
          type: boolean
          example: true
          description: whether additional synths cannot be minted
        synth_supply_remaining:
          type: string
          example: "123456"
          description: the amount of synth supply remaining before the current max supply is reached
        derived_depth_bps:
          type: string
          example: "123456"
          description: the depth of the derived virtual pool relative to L1 pool (in basis points)
        trading_halted:
          type: boolean
          example: false
          description: indicates if the pool can be used for swaps
        volume_asset:
          type: string
          example: "123456"
          description: 24h volume in asset
        volume_rune:
          type: string
          example: "123456"
          description: 24h volume in rune
...
    PoolsResponse:
      type: array
      items:
        $ref: "#/components/schemas/Pool"
...
    NetworkResponse:
      type: object
      required:
        # Since the querier uses an openapi. object, these will be displayed in alphabetical order,
        # so the schema should also be in alphabetical order.
        - bond_reward_rune
        - total_bond_units
        - available_pools_rune
        - vaults_liquidity_rune
        - effective_security_bond
        - total_reserve
        - vaults_migrating
        - gas_spent_rune
        - gas_withheld_rune
        - native_outbound_fee_rune
        - native_tx_fee_rune
        - tns_register_fee_rune
        - tns_fee_per_block_rune
        - rune_price_in_tor
        - tor_price_in_rune
        - tor_price_halted
      properties:
        bond_reward_rune:
          type: string
          example: "857134475040"
          description: total amount of RUNE awarded to node operators
        total_bond_units:
          type: string
          example: "0"
          description: total bonded RUNE
        available_pools_rune:
          type: string
          example: "4785723312627752"
          description: RUNE in Available pools (equal in value to the Assets in those pools)
        vaults_liquidity_rune:
          type: string
          example: "5792105087034476"
          description: RUNE value of Layer 1 Assets in vaults
        effective_security_bond:
          type: string
          example: "0"
          description: effective security bond used to determine maximum pooled RUNE
        total_reserve:
          type: string
          example: "21999180112172346"
          description: total reserve RUNE
        vaults_migrating:
          type: boolean
          example: false
          description: Returns true if there exist RetiringVaults which have not finished migrating funds to new ActiveVaults
        gas_spent_rune:
          type: string
          example: "1000000000"
          description: Sum of the gas the network has spent to send outbounds
        gas_withheld_rune:
          type: string
          example: "1500000000"
          description: Sum of the gas withheld from users to cover outbound gas
        outbound_fee_multiplier:
          type: string
          example: "15000"
          description: Current outbound fee multiplier, in basis points
        native_outbound_fee_rune:
          type: string
          example: "100000000"
          description: the outbound transaction fee in rune, converted from the NativeOutboundFeeUSD mimir (after USD fees are enabled)
        native_tx_fee_rune:
          type: string
          example: "100000000"
          description: the native transaction fee in rune, converted from the NativeTransactionFeeUSD mimir (after USD fees are enabled)
        tns_register_fee_rune:
          type: string
          example: "1000000000"
          description: the thorname register fee in rune, converted from the TNSRegisterFeeUSD mimir (after USD fees are enabled)
        tns_fee_per_block_rune:
          type: string
          example: "20"
          description: the thorname fee per block in rune, converted from the TNSFeePerBlockUSD mimir (after USD fees are enabled)
        rune_price_in_tor:
          type: string
          example: "10"
          description: the rune price in tor
        tor_price_in_rune:
          type: string
          example: "10"
          description: the tor price in rune
        tor_price_halted:
          type: boolean
          example: false
          description: indicates if all anchor chains are halted (true), or at least one anchor chain is available (false)
...